7.7 승미 푸쉬~~~끝나간다앗

+++ front/final_project/src/components/community/MyList.vue
@@ -1,15 +1,16 @@
 <template>
 
     <!-- 전체리스트 -> munnum이 같은 리스트 -> member 검색(컨트롤러)확인 후 작업 -->
-
+  <div id="container">
       <div class="mylist" v-for="(mycommlist, index) in commAllList" :key="index">
 
+        <!-- box1 start -->
         <div id="box1">
           <!-- 프사 & 닉네임 -->
           <div class="item-1">
-            <span><img style="margin-right: 8px; border-radius:50%; width: 30px; height: 30px;"
+            <span><img style="margin-left: 5px; border-radius:50%; width: 20px; height: 20px;"
               :src="'http://localhost:8081/members/imgs/' + mycommlist.memnum.memnum"></span>
-            <span style="margin-top: 8px;">
+            <span style="margin-top: 6px; margin-left: 7px; font-size: 0.5em; font-weight: bold; ">
               {{ mycommlist.memnum.nickname }}</span>
           </div>
 
@@ -19,30 +20,41 @@
             </button>
           </span>
         </div>
+        <!-- box1 End -->
 
-        <div class="imgBox">
+        <!-- box2 start -->
+        <div id="box2">
           <img class="img1" :src="'http://localhost:8081/ocommunity/img/' + mycommlist.commnum + '/' + 1">
-          <span v-if="mycommlist.img2 != undefined">
-            <img class="img1" :src="'http://localhost:8081/ocommunity/img/' + mycommlist.commnum + '/' + 2">
-          </span>
-          <span v-if="mycommlist.img3 != undefined">
-            <img class="img1" :src="'http://localhost:8081/ocommunity/img/' + mycommlist.commnum + '/' + 3">
-          </span>
-        </div>
 
-        <div class="box2">
-          <div class="myTag">{{ mycommlist.tag }}</div>
+          <div v-if="mycommlist.img2 != undefined">
+            <img class="img1" :src="'http://localhost:8081/ocommunity/img/' + mycommlist.commnum + '/' + 2">
+          </div>
+
+          <div v-if="mycommlist.img3 != undefined">
+            <img class="img1" :src="'http://localhost:8081/ocommunity/img/' + mycommlist.commnum + '/' + 3">
+          </div>
         </div>
-        <div class="box3">
-          <div class="likeBtn"> <!--좋아요 버튼-->
+        <!-- box2 End -->
+
+        <!-- box3 start -->
+        <div id="box3">
+          <div class="likeBtn">
 						<button class="markbtn1" @click="pushLike(mycommlist.commnum)">
-							<span class="material-symbols-outlined" :style="{'color' : mycommlist.chklike ? '#f15746' : 'lightslategray'}"> favorite</span>
+							<span class="material-symbols-outlined" 
+                :style="{'color' : mycommlist.chklike ? '#f15746' : 'lightslategray'}"> favorite</span>
 						</button>
 					</div>
           <div class="likeCount">{{ mycommlist.btnlike }}명이 좋아합니다.</div>
         </div>
-      </div>
+        <!-- box3 End -->
 
+        <!-- box4 start -->
+        <div id="box4">
+          <div class="box4-item">{{ mycommlist.tag }}</div>
+        </div>
+        <!-- box4 start -->
+      </div>
+    </div>
 </template>
 <script>
 export default {
@@ -66,15 +78,20 @@ export default {
   methods: {
     delPost(commnum) {
 			const self = this;
-			self.$axios.delete('http://localhost:8081/ocommunity/' + commnum)
-      .then(function (res) {
-        if (res.status === 200) {
-          alert('게시글이 삭제되었습니다.')
-          location.reload();
-        } else {
-          alert('에러코드: ' + res.status)
-        }
-      })
+      let check = confirm('정말 삭제하시겠습니까?');
+      if(check){
+        self.$axios.delete('http://localhost:8081/ocommunity/' + commnum)
+        .then(function (res) {
+          if (res.status === 200) {
+            alert('게시글이 삭제되었습니다.')
+            location.reload();
+          } else {
+            alert('에러코드: ' + res.status)
+          }
+        })
+      } else {
+        alert('삭제가 취소되었습니다.')
+      }
 		},
     pushLike(commnum){
 			let self = this;
@@ -94,25 +111,35 @@ export default {
   }
 }
 </script>
-<style scoped>
 
-.backBtn {
-  background-color:transparent;
-	border: none;
+<style scoped>
+@font-face {
+  font-family: 'PyeongChang-Regular';
+  src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2206-02@1.0/PyeongChang-Regular.woff2') format('woff2');
+  font-weight: normal;
+  font-style: normal;
 }
-.box1 {
-  border: #336399 solid 2px;
-  max-width: 48%;
-	/* margin: auto; */
-  margin-top: 2%;
-  margin-left: auto;
-  margin-right: auto;
+
+#container {
+  font-family: 'PyeongChang-Regular';
+  font-weight: 400;
+	display: grid;
+	grid-template-columns: repeat(auto-fill, minmax(25%, auto));
+  gap: 15px;
+}
+#box1 {
+	border-top-left-radius: 5px;
+	border-top-right-radius: 5px;
+	background-color: #c1f2ca2a;
+	max-width: 601px;
+	margin-top: 8%;
+	margin-left: auto;
+	margin-right: auto;
 	display: flex;
 	flex-direction: row;
 	justify-content: space-between;
 }
-
-.name-box {
+.item-1 {
 	margin-block-start: auto;
 	margin-left: 10px;
 	margin-top: 6px;
@@ -122,46 +149,56 @@ export default {
 .item-Btn {
 	background-color:transparent;
 	border: none;
-	font-family: sans-serif;
-	margin-top: 8px;
+	margin-top: 5px;
+	color: #f15746;
 }
-.imgBox {
-	border: #336399 solid 2px;
-	max-width: 48%;
+#box2 {
+	max-width: 45%;
 	margin: auto;
 	display: flex;
 	justify-content: center;
 }
 .img1 {
-  width: 200px;
-  height: 250px;
+  width: 80px;
+  height: 120px;
+  transition: transform 0.3s 
 }
-.box2 {
-  border: #336399 solid 2px;
-	max-width: 48%;
-	margin: auto;
-	display: flex;
+.img1:hover {
+	transform: scale(1.08);  
 }
-.myTag {
-  margin-left: 10px;
-	margin-top: 12px;
-	margin-bottom: 12px;
-}
-
-.box3 {
-  border: #336399 solid 2px;
-	max-width: 48%;
+#box3 {
+	background: #fff;
+	max-width: 601px;
 	margin: auto;
 	display: flex;
 	flex-direction: row;
 }
-.likeCount {
-	margin-top: 8px;
-}
 .markbtn1{
-	margin-top: 3px;
+	/* margin-top: 3px; */
 	background-color: transparent;
 	border: none; 
 	cursor: pointer;
+  margin-left: 3px;
 }
+.likeCount {
+	margin-top: 6px;
+	font-size: 0.7em;
+	font-weight: bold;
+}
+#box4 {
+	border-bottom-right-radius: 5px;
+	border-bottom-left-radius: 5px;
+	max-width: 601px;
+	margin: auto;
+	display: flex;
+	box-shadow: 16px 16px 16px rgba(195, 192, 192, 0.345);
+}
+.box4-item {
+	margin-left: 10px;
+	margin-top: 8px;
+	margin-bottom: 12px;
+	font-size: 0.6em;
+	font-weight: bold;
+}
+
 </style>
